+++ domains/SignIn/index.js
@@ -20,8 +20,7 @@ import { retrieveSignInFunction } from '../../services/parse/auth';
 import FormInput from '../../components/FormInput';
 import LanguagePicker from '../../components/LanguagePicker';
 import CredentialsModal from './CredentialsModal';
-import { storeData, getData } from '../../modules/async-storage';
-import * as Network from 'expo-network';
+import { storeData, getData, deleteData } from '../../modules/async-storage';
 
 import I18n from '../../modules/i18n';
 
@@ -68,7 +67,10 @@ const SignIn = ({ navigation }) => {
       [
         {
           text: 'Yes',
-          onPress: () => storeData(values, 'credentials')
+          onPress: () => {
+            storeData(values, 'credentials');
+            navigation.navigate('StorePincode');
+          }
         },
         { text: 'No', style: 'cancel' },
       ],
@@ -90,6 +92,9 @@ const SignIn = ({ navigation }) => {
     const { isConnected } = status;
     return isConnected;
   }
+  const deleteCreds = () => {
+    deleteData('credentials');
+  };
 
   return (
     <SafeAreaView style={{ marginTop: 20 }}>
@@ -191,10 +196,12 @@ const SignIn = ({ navigation }) => {
               formikProps={formikProps}
               user={user}
               setModalVisible={setModalVisible}
+              navigation={navigation}
             />
           </>
         )}
       </Formik>
+      <Button onPress={deleteCreds}>Delete Credentials</Button>
     </SafeAreaView>
   );
 };